benchmarks: merge cachegen_xpu_e2e into test_cachegen and whitelist GPU backends
The xpu_e2e file was a superset of test_cachegen.py: same KV-blob /
serde construction logic, but it benchmarked both encode and decode
(plus a compression-ratio print), while test_cachegen.py only had a
decode bench (encode was commented out). It also wasn't actually
XPU-specific -- it used torch_device_type, so CUDA could run it too.

Merge encode + compression-ratio print into test_cachegen.py, unify the
chunk_size parameter set to [64, 128, 256, 768], and delete the
duplicate xpu_e2e file. Switch the skip gate from is_available() to an
explicit cuda/xpu whitelist so HPU / CPU-only CI / future backends
without a CacheGen kernel are skipped automatically.

Verified on Arc Pro B60 (XPU): all 8 benchmarks (encode/decode x
4 chunk sizes) pass with real timings.
